inheritance

+++ chap08_inheritance_lecture_source/src/main/java/com/ohgiraffer/section02/car/Car.java
@@ -1,5 +1,10 @@
 package main.java.com.ohgiraffer.section02.car;
 
+// 자동차가 공통적으로 가지고 있는 속성을 만들어준다.
+/*
+* 자동차는 공통적으로
+* brand명과 price를 갖는다.
+* */
 public class Car {
     // 현대차
     private final String brand;
@@ -10,7 +15,7 @@ public class Car {
         this.price = price;
     }
 
-    public int getPrice(){
+    public double getPrice(){
         return this.price;
     }
